+++ package.json
@@ -13,6 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "async": "^2.6.1",
     "axios": "^0.18.0",
     "busboy": "^0.2.14",
     "http-proxy": "^1.17.0",
